backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pymongo
python-dotenv
bcrypt
